Partial fix for 180° turns (issue #3)
The other part is to advance the snake only by entire pixels at a time.
You can still trigger the bug if you are going right, press up/down,
"advance" 0.3333 pixels then press left, and the floor of your new
location equals the floor of the old. This is because using a position
variable with more precision than our collision-detecting routine is
silly :)