Revert "Turn on Closure Compiler optimizations"
This reverts commit a62bb688f2fa7b8be33e0744fe0194e91bae4c77.

Much as I like the smaller file size, :optimizations :advanced does
not play well with incremental compilation — compiles after the first
appear to break randomly. It also will not work with a browser REPL
for me, yielding an undefined variable error in the JS console. It now
appears that :optimizations :advanced must only be used when ready to
deploy JS to production, and even then I'll have to clean before every
build.